+++ docs/changes.md
@@ -3,6 +3,7 @@
 2023/10/25
   * 增加函数tk\_iostream\_serial\_create\_ex
   * 修复编译报错的问题(感谢智明提供补丁)
+  * 增加函数str\_append\_lowercase/str\_append\_uppercase
 
 2023/10/24
   * 完善stream。